risk errors moved to domain
