+++ crates/driver/src/infra/solver/dto/solution.rs
@@ -200,7 +200,7 @@ impl Solutions {
                         Score::Solver(score) => competition::solution::SolverScore::Solver(score),
                         Score::RiskAdjusted(success_probability) => {
                             competition::solution::SolverScore::RiskAdjusted(
-                                competition::score::SuccessProbability(success_probability),
+                                competition::score::risk::SuccessProbability(success_probability),
                             )
                         }
                     },
